Replace RunLoop#begin/end with RunLoop#invoke
Instead of calling begin/end to create a run loop, you now invoke a function via
a call to RunLoop#invoke. This allows errors to be caught, making applications
much more robust as you never end up in the state where the RunLoop thinks it's
at depth > 0, but isn't. It also allows errors to be easily posted to the server
for logging.