MenuView: Take focus on mousemove, not mouseover.
When a DOM element is inserted on screen, Firefox will immediately fire the
mouseover event if the cursor happens to be over where it appears on the screen.
This can cause an apparently random item in the menu to be initially focussed.
Instead of mouseover, we now use mousemove, so that it only focuses when the
user starts using the mouse again.