Use common name for object init mixin argument.
Most constructors take a single argument: an object of methods and properties
which is mixed in to the new object before it is initialised (essentially
equivalent to configuring default values and/or creating an anonymous subclass).

This argument was called either `options` or `config`. Now renamed everywhere to
`mixin` for consistency and clarity.