Handle transitionend not firing.
If asked to animate from x -> y then very soon after from y -> x again, the
browser may never have actually drawn a frame in the first transition, so may
not fire the transitionend event. To ensure the animation end is detected
properly, a timeout is now set for after the animation should have finished. If
the transitionend event doesn't fire, the timeout will be triggered and cleanup.