Avoid race condition refreshing queries.
The OBSOLETE flag is now removed when the query update refresh is initiated,
rather than when it completes. This means if you make a change that causes the
OBSOLETE flag to be set whilst the query is being fetched, that flag will remain
after the fetch completes (since we don't know whether the fetch completed
before or after the event that caused the flag to be set).