Reverse preemptive change after server update.
If you apply a preemptive update to a windowed remote query, the query is now
marked as being in the DIRTY state. This flag is removed when an update is
fetched from the server. When an update is applied via
O.WindowedRemoteQuery#sourceDidFetchUpdate, if it says nothing has changed but
there are preemptive updates currenty applied the preemptive changes must be
wrong, so they are reversed. If further changes have been made whilst the update
was being fetched from the server, the DIRTY flag will have been set again and
the preemptive changes will not be undone (providing the changes that *are*
fetched are a subset of those applied); the next updates fetch will find if
these need to be undone.