Use fallback iOS scroll solution when pos=relative.
The preferred solution for ensuring scroll views always scroll even when at the
end on iOS is to use a wrapping div. However, this doesn't work if the
ScrollView is relatively rather than absolutely positioned, so in this case we
fall back to the previous method of resetting the scroll position to 1px off
the ends after the scroll has finished.