Wait for list to become ready in SingleSelectionController
When first initialising an application, the data may not yet be loaded for the
initial list. However, we may have a record we would like to set based on the
URL. We must wait for the data to load before we can determine if the record
still exists in this view.