Format KB shortcuts according to platform conventions.
On a Mac, the symbols are normally used, e.g. "⌘S". On Windows/Linux, it's
normally written like "Ctrl-S". A new function: O.formatKeyForPlatform now
handles this formatting automatically.

In addition, a new modifier, "cmd-", may be used either with this function or
with the global keyboard shortcuts object to automatically map to the cmd key
on Mac and the ctrl key on Windows/Linux.