MenuView: Focus first match when filter changes.
When the filter changes in a MenuView, we keep focussing the currently focussed
option so long as it matches. If it no longer matches however, we should focus
the first match in the list, rather than the first match after the previously
focussed one in the full list, which appears random.