Fix list scroll adjustment when animated items.
When an item is inserted above the currently visible section of a
ProgressiveListView, we adjust the scroll position of the containing scroll
view so that the first visible item remains the same. However, if the items are
animated, instead of making it appear like there's no change, it would "jump"
up an item, then scroll back down to the original position. This commit fixes
the jump by redrawing the new positions before any adjustments for animations
take place.