eslint no-use-before-define: fix some bugs
This one actually found a few genuine bugs in my ES6 module translation,
specifically ones of the form `var foo = foo;`. I allowed these to
happen by accident and fixed some of them when I figured out what was
going on, but forgot to check exhaustively for them later on. Thanks,
eslint, for incidentally reminding me.

The other ones aren’t bugs while `var` is being used, but would be if
`let` was used instead, which I want to be doing soon. So it’s a matter
of future-proofing the coding style.