WIP: use ES6 methods more directly
At this point I declare that Overture should assume an ES6 baseline
spanning the intersection of all that is supported by all current major
browsers, desktop and mobile, plus es6-shim for browsers that don’t
support `class` (most specifically, IE 11).

Connected tangentially with this is the slightly more subjective stuff
of phasing out global prototype pollution (because it’s an undesirable
practice and because new ES features allow us to replace various of
these things we’re injecting with other techniques without much pain).

Steps one and two of this, included in this commit are:

- Deprecating `Function#implement`, `Function#extend` and `O.extend` in
  favour of `O.mixin` and/or `Object.assign`;

- Removing other ES6 methods (Overture can assume an ES6 baseline now):
  - `Object.values`
  - `String#repeat` (our polyfill was simplistic anyway)

But getting back to that ES6 baseline and es6-shim, I think we should
experiment with generating two versions of the code:

- Feature detect on support for the `class` keyword (primarily for IE
  11), ES5 syntax via Bublé/Babel, with es6-shim or core-js or something
  like that to provide ES6 methods like `Object.assign`, `Object.values`
  and `String#repeat`.

- ES6 syntax, for everyone else.

We can start using things like `Map` and `Set` too, which should subsume
most or all of the uses of `Array#include` and `Array#erase` (which
smell like poor man’s `Set`) so we can ditch them.