Remove duplication of SVG in module bundles
*.svg was being picked up as both a font and an image, leading to it
being written to the bundle twice: once as a base64-encoded URL, and
then again (contrary to the documentation, argh!) as a plain
stringification of the SVG contents. The latter overrides the former, so
there’s no point in the former being there.

(Plus no one uses SVG fonts.)

This should reduce the SVG portion of bundle sizes substantially.