Replace fs.writeFile to avoid deprecation warnings
fs.writeFile without a callback is deprecated in Node 7. All usage of it
is herein replaced with fs.writeFileSync; this has the potential to be a
trivial performance loss, but… do we even care? Nah, not this time. And
it’s too much work to buy into the async lifestyle altogether for these
scripts. Let’s just silence these annoying warnings for now and work on
improving the build tools for the Overture ecosystem later.