Use a <span> not a <label> for TextView layer tag
The label for a text view is rarely put inside the view, unlike other controls.
You only want one <label> associated with the input, so let's switch to a <span>
as the default tag, and then you can create your labels externally without
having to override the layerTag all the time.