Complain about `O.Class({ Extends: undefined })`
Specifically, we experienced this happening with circular imports due to
ButtonView depending on MenuView, and so ClearSearchButtonView ended up
coming before ButtonView, leaving a bad class definition. Although this
is strictly legal and could conceivably have a legitimate use, said use
is so utterly convoluted that we decided to forbid it with an immediate
exception—much easier to debug than a later uncontrolled explosion.