Fix an O.Object/window.Object confusion
This was a regression in the modern Overture project.

Only one mention of Object was actually broken, in SelectionController.
The other changes are preventative and for consistency.

The other two that would have been issues (in Store and View) I caught
in the first place and had imported O.Object under the name OObject.
We’re going with the uniform convention “Obj” now, so I updated all
O.Object imports to match.