Improve handling of non-200 responses in loader
Very rarely with FastMail or Topicbox you could wind up with the module
loader caching a 404 if the code disappeared for some reason, such as a
mistaken removal of files from the production server. Unsurprisingly,
this 404 response text wasn’t the JavaScript desired, so the app would
be broken until the bootstrap changed the source for that module. For
the most part this meant it didn’t matter, because the only reason a
file would disappear was changes to the module and corresponding
bootstrap entry, so reloading it would clear the bad localStorage entry,
but in cases where something did go wrong with the server briefly so
that a 404 or similar was given *without* there being changes to the
module, the user was left with the app just broken, with clearing the
localStorage being the only solution.

I’m not fond of the `confirm` I used here, but I think it’s better than
nothing given the rest of our module loading infrastructure. It gives
the user an indication something has gone wrong, at the least. I
designed it so it is easily overridden for a better, app-specific UI.

This only takes care of the `!useScriptTag` case. In the much rarer
`useScriptTag` case, the browser knows only to execute a 200 response.
We don’t do anything about errors, however, so in case of 404s the user
will just be left hanging at present. We could do something about it
but given its rareness and the extra code required… meh.