SVG support for Element
O.Element.create can now create SVG elements (which require the SVG
namespace); this allows you to use `el( 'path' )`, `layerTag: 'svg'` and
the likes.

Also Element#{get, set} now abstract SVGAnimatedString out of the way
for the sake of the className property. We don’t need the fanciness that
SVGAnimatedString provides; we’re not using SMIL.

For “completeness” we’d need to add MathML as well, but we don’t need it
at FastMail at all and so I am disinclined to add the weight of listing
all the MathML tags.