Make View#id behave more strictly as documented
The documentation of O.View#id said:

> The id of the view. Automatically assigned if not overridden. Will be
> set as the id of the underlying DOM node.

This was only partially true: it would always automatically assign an
ID, but if you passed in `id` in the mixin parameters that would
override the automatically assigned ID.

But I want to be able to do things like define `id` as a computed
property in the prototype. Care must still be taken to ensure that this
ID can’t change later, but this is a legitimate scenario. (Frankly, I
don’t see why ID shouldn’t be allowed to change after view rendering,
but that’s another matter.)