[Breaking change] Rejig active view registration
`O.ViewEventsController._activeViews` has just evolved!

It’s been made public as `O.activeViews`, and has the new superpower
that views with manually-specified IDs are now **observable**: suppose
you are waiting for a view with ID `foo` to enter the screen: you could
modify the definition of `foo` to do set some other property when it
enters the document (this is often going to be a better way to do it),
mor you can just bind things to “the view with ID `foo`” like this:
`bind( activeViews, 'foo' )`. I’ll leave it to your imagination to
decide why you might want to do this. I know why I want it (and
implemented what was effectively this, but outside of Overture). This is
not something you should use terribly often, but it’s very useful when
you do want it. Without it, you’d be stuck making more intrusive changes
to lots of views, or using a MutationObserver and watching the *DOM*
until your element with a given ID is added to the document.

Breaking changes:

- Things pertaining to active views were moved out of
  `O.ViewEventsController` because they were only tangentially related
  to it. (I think they were only there because `ViewEventsController` is
  the first thing that needed something like `getViewFromNode`.)

  - **`getViewFromNode` has been moved to `O.getViewFromNode`.**
    This is the main one that may break live code.

  - The `registerActiveView` and `deregisterActiveView` methods have
    disappeared, inlined into the sole methods that used them. They were
    only suitable for internal use from `O.View` anyway, so this
    shouldn’t break any real code.

  - As noted at the start, the `_activeViews` private property has been
    made public, being exposed as `O.activeViews` with certain added
    functionality.

- There is a downright minuscule possibility that the following change
  will break things: View#id was formerly always set before calling
  O.Object#init (even if the mixin arguments contained an `id` which
  would then override that automatic ID), but an automatic ID is now
  only assigned *after* the super `init` call.