PopOverView: fix misunderstood calloutSize prop
In commit 496ed1afc495c296da30f17c0cd83d9a14b3b852 (PopOverView: extract
callout size magic number) I misread the meaning of the magic number
ten. It’s nothing to do with the callout size, it’s to do with the
margin that is maintained around the popover, inside the parent view.

Given that, I don’t think it’s particularly useful to override it, but
magic numbers should still be extracted, so I redefined it as a constant
PARENT_MARGIN. The property PopOverView#calloutSize is gone again.