Add support for separate "visible" query in SelectionController
The "content" property should be the query with the full set of available
options. If something is removed from here, it is automatically removed from
the selection. If this changes, the selection is reset.

The "visible" property should be the list currently displayed to the user
(which may be a subset of "content" if the user is filtering). This is used
when selecting indexes/ranges to match what the indexes of items in a ListView
being displayed to the user.