Deprecate Element#{addClass,removeClass,hasClass}
Our browser support base supports at least the basic version of
Element#classList (where each method takes only one argument) without
any polyfills; it’s nicer to use that directly.

Actually, I should point out that classList isn’t available on SVG
elements in IE 11. You’ll need a polyfill if you need to work with that.
Hmm, I guess that is something to be aware of if I try changing View
from working with strings (View#className) to using a list (not that I
have tried it yet or am likely to any time soon, and I don’t know if
it’s a good idea).

`O.Element` is a namespace; it should really have a lowercase first
letter, and avoid a collision with the browser’s global type `Element`.
Maybe we should rename it to `O.dom`. Or maybe pull it apart altogether.
Just as addClass, removeClass and hasClass did, setStyle and setStyles
also return `this`, which is very icky for a namespace.