Integrate native Promise with the RunLoop
This is not going to be especially brilliant for performance, but it’s
basically something that would already have been required anyway.

I’ve verified that async functions are using the global `Promise` (thus
our `O.Promise`) in both Firefox and Chrome, so with this it’ll be OK if
we ever want to write async functions—though we’ll have to make sure
that `self.Promise` is replaced before the async function is *called*,
or we’ll have a non-RunLoop-enabled promise on our hands. For now, I’m
content replacing `self.Promise` in `Overture.js`.