Promise integration: fix transpilation breakage
It worked fine with class syntax, but when transpiled to function syntax
it fell apart, because JavaScript isn’t exactly a flawless beastie and
class syntax is subtly different from the generally equivalent function
syntax, and `new` is downright crazy. So I rewrote it in a form that
works, based on function syntax, because we can’t use class syntax for
all our browsers.