RunLoop#invoke: return the function’s return value
Also Function#invokeInRunLoop now returns the wrapped function’s return
value instead of `undefined`.

Together, these fix the `Promise` integration. (The first commit with it
was completely broken, the second was only broken if you depended on the
return value of a `then()` call, maybe this third take is sound!)

This is an obscure breaking change that is unlikely to affect real code.

If using this, bear in mind that the run loop queues will be flushed
between the completion of the function and the return of its return
value, so if you’re trying to do the wrong thing with it, that value
could be stale.