Fix up broken Promise#{then,catch} behaviour
This came up because of a dodgy user script triggering an unexpected but
perfectly legal code path.

onFulfilled and onRejected can both be non-function values.

And Promise#catch internally calls Promise#then with the first argument
undefined. So my reimplementation of it was simply broken. Clearly no
promise was rejected before this! I’m amazed at my foolishness and lack
of testing in this whole promises issue. It’s remarkable what the lack
of a proper test harness has done to me.

Anyway, Promise#then now only wraps each argument in invokeInRunLoop iff
it is a function. If it’s not, it just passes it through. And the catch
function is removed, because it’s not necessary after all.