Fix a subtle regression in TimeZone’s getRule
I introduced this regression in the modernisation, where I changed it
from `var` to `let`, and (critically) shifted it into the innermost
scope where the variable was used. The combination of these two changes
(they’d be fine individually) meant that the variable was reset to
`undefined` on each iteration of the loop, which was *not* what the
doctor ordered.

I’m too lazy to think through the implications of the broken behaviour.
Time zones are evil, I’m willing to just be happy that it’s fixed, and
sigh to myself about the lack of a test suite.

I picked it up by feeding Overture through https://deepscan.io/.
I’m pretty impressed with it.

I’m a little concerned that I might have done this elsewhere, but I
think that DeepScan would pick any other such cases up.