Remove unnecessary null checking
In View: the null checking of `layer` is unnecessary, because
isInDocument has already ensured that it will not be null; but more to
the point, `layer.offsetTop` and `layer.offsetLeft` in the previous
lines would have already made it blow up if layer *were* null.

In ScrollView: before modernisation, this was `if ( NS.Tap )`. This was
rendered unnecessary by Rollup’s proper dependency management.

Another couple of thingies pointed out by DeepScan.