Remove superfluous property-setting on Animation
Contrary to popular belief (and DeepScan’s opinion), these *did*
actually serve a purpose: they ensured a property was set on the
instance, rather than just leaving it to the prototype.

`Animation#duration` and `Animation#ease` are both supposed to be
readonly on an `Animation` instance, so there’s no performance benefit
to assigning it during construction—either they’ll be set later in the
constructor, or they never *needed* to be set. This is merely an
interesting curio that came about from a code style alteration long ago.
I’m inclined to agree with DeepScan that it’s undesirable.

If you write dreadful enough code, this is actually a minor breaking
change: formerly, if you changed `Animation.prototype.duration` to 1000,
any Animation instances created before that change (and without
`duration` in their mixin) would have their duration set to 300 (the
Overture default) instead of 1000. Now, they’ll change silently
underneath you to 1000, possibly making your cat eat your laundry or
your dog your homework if the animation is running. But seriously, you
shouldn’t be changing things on prototypes like that. It’s nasty. If you
want to do that sort of thing, subclass the type or come up with some
other way of expressing it, please. Just because you can do it because
this is JavaScript doesn’t mean that you should. Assigning to any object
in the `O` namespace in client code is *wrong*, say I. (Yes, the
FastMail and Topicbox frontends both do this in a very few situations,
and I wish they didn’t.)