Transform Cmd- keyboard shortcuts appropriately
With event.key we get a logical result back, and this differs if the Ctrl vs
Meta key is held down. Quick fix, if the shortcut includes Shift- and we're not
on Mac (where we use Meta, not Ctrl), transform the shortcut letter to
upper-case (yes, yes, for a very English-orientated version of upper-case).