PopOverView: add an option to resist hiding
This allows you to make a popover modal, requiring more deliberate
action to close it—e.g. a “close” button inside the popover, rather than
just clicking outside it or pressing Escape. (The particular case we
need it for is a login popover that, once begun, should not be closed
lightly.)

I’m not really happy with the shape of PopOverView and how options are
handled in it and how that meant I had to allow resistHiding to be a
function, but this is what we have for now.

(Well, I considered making it a getter, because it’s OK for us to depend
on descriptors now, but I decided not to introduce a new pattern for
this!)