localise(): support objects and macros together
When at least one of the formatting arguments is an object,
there formerly existed a limitation that macros would not work.
This limitation finally annoyed me enough that I’ve fixed it.

(It was causing trouble in generic code that should have been able to
deal with strings or objects, but also needed cardinality checking in
the same string; while handed strings, it would be happy, but once
handed an object, it wouldn’t process the `[*2,_2,…]` part.)