Improve key support
This fixes a regression in Firefox < 37 (not sure about other browsers
we may care about) from when we switched from the physical to logical
models a month or two back, where the arrow keys would no longer work.

It also fixes Escape and the arrow keys on iOS Safari, where I
discovered today that Apple have made a bit of a mess of things. Filed a
bug report about it, so hopefully it’ll be fixed for everyone soon. In
the mean time, at least it’ll work with us.