loc(): cope better with extraneous square brackets
14cd45e197c525e30f831b1f5be538dcc794b511 regressed this unintentionally.

This is a corner case that we want to work properly in Topicbox: square
brackets inside square brackets, like `foo[[_1]]`, where _1 is an
element. It’s worth noting that if _1 is a string, an exception will be
thrown! It would probably be better in such a case to add a couple of
extra arguments, _2 and _3, which are the square brackets, [ and ], but
for now I just want to fix this regression.