Show floating toolbar for keyboard selection too
A RichTextView with a floating toolbar was only showing the toolbar when
the mouse was used to make a selection; this makes it respond to a
keyboard selection as well.

The event handler is very lightweight (a couple of O.Object#get calls to
non-computed properties, and a couple of comparisons), so no performance
issue should come of it even when you type at speed.