Tweak date parsing
Allow date like "2w" without +/- prefix (presume negative for compatibility
with Perl date parser). Also allow a space after the number, e.g. "2 weeks".

The parse search order is reprioritised so a lone number is not greedily parsed
as an hour or day, to allow for the space after the number in adjust dates.