Make PopOver#options public and keep anchored
The pop over may slide left/right if positioned to the top/bottom of
a view, or slide up/down if positioned to the side. This means it is
never moved over the view it is anchored to.

The contents view must make sure is not too tall/wide to fit given the
space constraints imposed by this.