Change semantics of Record status OBSOLETE
This now means the server has indicated data has changed in this record, and no
new data has been loaded into the store since, i.e. the data in the store is
currently out of date. Previously this flag was cleared when you requested the updated data; now it is only cleared when the updated data loads.

This fixes a problem where once a record was marked OBSOLETE, unsolicited data
for it would be ignored so you could end up double fetching the record.