Sort render queue before dispatching
We sort by view depth, and views before other items (which are normally
DOM nodes that are being updated directly). This means that parents are
always redrawn before their children, which is often more efficient, for
example if the child is removed by the parent redrawing, we haven't
bothered to redraw it first.