Store: handle hasDataFor more like JMAP
It’s nicer for us to accept it as an array here rather than taking an
object—we can reasonably assume it’s easier for the caller to pass it in
as an array (either manually specified, or more likely from a JMAP
session object), and then we can do the transformation in one place.

Removed support for it being null, too, since it’s mandatory in JMAP.

Consequently, I also removed the dummy default account that wouldn’t
actually be used for anything in practice; now, `store.addAccount()`
must be called manually before you can use the store.