RecordResult: use a symbol for HANDLE_ALL_ERRORS
We haven’t used any symbols yet, but we’re safe to; our browser baseline
says “Object.values, or core-js” which covers symbols.

Using object identity on an array was decidedly unpleasant there; it’s
fine for HANDLE_NO_ERRORS to be an empty array, because that’s
semantically what it *is* (hence leaving it be), but HANDLE_ALL_ERRORS
is instead a sentinel value. You’re not likely to ever need to debug it,
but distinguishing between two empty arrays for completely the opposite
behaviour is… yeah.