Fix bug accidental property override
The OptionsListView#focused property was blatting the new ListView#focused
property, occasionally causing errors to be thrown. I *do* like languages that
force you to declare when you deliberately want to override an existing property
in a parent; once we have property annotations we could probably implement this
in Overture. Hmm…