Add ability to ignore server state changes
A client doing a large batch of changes may split these up into several commits
so as not to overload the server. While doing this, side effect changes may
cause state strings to change and so normally the client would continually
fetch the changes. It may be quicker to wait and do this at the end though, so
now the client can set the ignoreServerState flag on an account in the store
while it is processing the changes, and then check once afterwards.