Router: ditch useHash in favour of more baseUrl
All browsers that we now care about support the History API, including
the popstate event. (IE supports it from 10 on, and we only care about
IE 11 now.) Due to this, we no longer need the hashchange fallback, and
so by tweaking how baseUrl works, to make it purely a prefix (thus, for
https: URLs, including the origin), we can drop useHash altogether.

useHash used to *almost* work on the file: protocol if you didn’t
include “#/” in the URL when you loaded the page—it’d simply think you
were on a very strange URL, and you might be able to recover from it.
Now, the router will fail to initialise. I consider this to be an
acceptable pain, when the alternative is greater complexity and
unnecessary bytes served (I’d prefer that the production bundle didn’t
support the 'file:' protocol at all, but our build toolchain isn’t in a
suitable shape to support that yet).

The rest of this commit message justifies removing the exeption
suppression around the history.pushState/history.replaceState calls.

I removed it because I don’t like seemingly-unnecessary exception
handling, and I haven’t had any trouble with it in Firefox in quite some
years. All situations where the spec says it may produce an exception,
we should care about; so I prefer also to stop masking such things. See
https://html.spec.whatwg.org/multipage/history.html#dom-history-pushstate
for the algorithm for history.pushState and history.replaceState. Let’s
look at the possible cases:

> 2. If document is not fully active, throw a "SecurityError"
>    DOMException.

I don’t believe there are any situations in which our document will not
be fully active.

> 5. Let serializedData be StructuredSerializeForStorage(data).
>    Rethrow any exceptions.

We always pass null for data, which serializes trivially according to
that algorithm.

> 7. If url is not null, then:
>
>    1. Parse url, relative to the relevant settings object of this
>       History object.
>
>    2. If that fails, then throw a "SecurityError" DOMException.

Provided baseUrl is set properly and the getUrlForEncodedState function
returns a sane value, this won’t be triggered. If it is triggered, it’s
a bug that we don’t *want* suppressed.

>    4. Compare newURL to document's URL. If any component of these two
>       URL records differ other than the path, query, and fragment
>       components, then throw a "SecurityError" DOMException.

If something causes us to be trying to change the scheme, host name or
port, something’s badly wrong, and we don’t want to suppress it.

>    5. If the origin of newURL is not same origin with the origin of
>       document, and either the path or query components of the two URL
>       records compared in the previous step differ, throw a
>       "SecurityError" DOMException. (This prevents sandboxed content
>       from spoofing other pages on the same origin.)

The main time this affects us is when serving from the file: scheme,
where the origin is probably an opaque origin that internally contains
the full path name. This is why you can, on file: URLs, modify the hash,
but not the path or query string. And again, if we’re somehow doing
something wrong with this, we don’t *want* to suppress it.

In summary: the try/catch around history.pushState and
history.replaceState is undesirable unless it’s for working around
innocuous browser bugs, and I have no evidence that such bugs yet exist.