Drop Router#mayGoBack in favour of dynamic routes
The purpose of mayGoBack was to block navigation while awaiting login,
but it seems to me that making Router#routes be dynamic is a better and
more general solution.

Route handlers should be idempotent. If they’re not, things may go
screwy—that was already potentially the case, but is so even more now.