Router: refactor routing, ditch currentPath
Renamed restoreStateFromUrl to restoreEncodedState, because its
argument is an encoded state, not a URL. I thought of calling it
restoreStateFromPath, but I’ll shortly be adding query string support,
and I’d only need to rename it again after that, because calling
“foo/bar?baz=quux” a path would only make us miserable.

(I’m not fond of the two meanings of “state” that we have going here: we
have our own routing state, and then there’s the History API’s notion of
state which we don’t use (nor does anyone else, as far as I know—it was
an interesting *idea* only), leading to Router#replaceState which means
“use history.replaceState next time instead of history.pushState”.)

Ditched currentPath because now that Router#routes is dynamic and
Router#mayGoBack is gone I can’t think of any situation in which it’s
necessary—just use location.href instead.

This might have a subtle effect on how using the hash works, so I
documented that. We don’t deliberately use the hash anywhere, so I think
we’re probably OK with the implications there. That’s also why I have
gone to almost no effort to fix the potential slight breakage.
(Certainly if running an Overture app on the file: scheme, regular hash
usage never had a chance of working!)