Router: support global query string parameters
I think this is a good design. It should be both more robust and more
flexible than what we use in FastMail and Topicbox.

The biggest blemish in the design to note is how getUrlForEncodedState
is not a pure function: I excused baseUrl since it’s designed to be set
once only, but _globalQueryStringPart is expressly not immutable now.
The end result is this scenario:

1. Have a known global query string parameter debug (property name also
   debug);
2. Load https://www.example.com/?debug=1: router.debug === '1'.
3. Render an internal link: https://www.example.com/foo?debug=1.
4. router.set( 'debug', null ): ?debug=1 disappears from the URL, but
   not from the already-rendered link.
5. Click on the link and have that encoded state restored: router.debug
   becomes '1' again.

See the TODO I left; we can make it better in theory with a new
observable property, but in practice I don’t think it’s a big deal (I
can only think of one place in existing code where I want to use this
functionality, and it’s right at the start, immediately after the router
is constructed, before any rendering begins, so it won’t trigger the
problem) and is not likely to be used (who’s actually going to observe
this extra obscure property?). So I think I’m sufficiently content to
just leave the bug.